Fix small fonts
The two big ancestry figures had fonts that Springer said were
less than 2 points, which obviously wasn't going to be readable
when printed. This hopefully fixes that, but I can't actually figure
out how to find out what font sizes are being used in the PDF.
I'm a little worried that figure 5.1 is still going to be a problem.